\name{Calc_Energy}
\alias{Calc_Energy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Calculating the energy of each level}
\description{
  loop over each level
}
\usage{
Calc_Energy(X, basis, K, av_norm)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{the input matrix }
  \item{basis}{basis matrix at certain level, please refer to function JTree\_Scores}
  \item{K}{number of vectors in final representation}
  \item{av_norm}{normalization term}
}

\value{
  \item{}{the score at the level}

}

\examples{



}

\keyword{ math }

