\name{treelet-package}
\alias{treelet-package}
\alias{treelet}
\docType{package}
\title{
Treelets - an adaptive multi-scale basis for high-dimensional,sparse and unordered data }
\description{
Treelets provides a novel construction of multi-scale bases that extends wavelets to non-smooth signals. It returns a multi-scale orthonormal basis, where the final computed basis functions are supported on nested clusters in a hierarchical tree. Both the tree and the basis, which are constructed simultaneously, reflect the internal structure of the data. 
}
\details{
\tabular{ll}{
Package: \tab treelet\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-02-10\cr
License: \tab GPL (>= 2) \cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Di Liu and Trent Gaugler

Maintainer: Trent Gaugler <gauglert@lafayette.edu> 

%~~ The author and/or maintainer of the package ~~
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
