\name{Calc_Energy}
\alias{Calc_Energy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A little function calculate the energy of each level}
\description{
  loop over each level
}
\usage{
Calc_Energy(X, basis, K, av_norm)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{The input matrix}
  \item{basis}{basis matrix at certain level, please refer to function JTree\_Scores}
  \item{K}{number of vectors in final representation}
  \item{av_norm}{normalization term}
}

\value{
  ~Describe the output
  \item{}{the score at the level}

}
\author{Ann Lee, Di Liu }

\examples{



}

\keyword{ math }

